+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HFP HF: send HF Indicator update only if enabled by AG
 
 ### Changed
+- HCI: hci_reserved_packet_buffer and higher layer functions asserts if packet buffer is free instead of return value
 - GAP: mutual authentication for legacy secure connections to prevent BIAS attacks was made optional with 
        ENABLE_MUTUAL_AUTHENTICATION_FOR_LEGACY_SECURE_CONNECTIONS. Not needed for default encryption key size of 16